implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\QuestionRequest;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -62,6 +63,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(Gate::denies('update-question', $question)) {
+            abort(403, "Acces Denied");
+        }
+        
         return view('questions.edit', compact('question'));
     }
 
@@ -86,7 +91,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(Gate::denies('delete-question', $question)) {
+            abort(403, "Access Denied");
+        }
+        
         $question->delete();
         return redirect()->route('questions.index')->with('success', "Your Question Has been Deleted");
+       
     }
 }
